updated homwork

+++ seb_b_q_12.py
@@ -1,7 +1,7 @@
 # 12. Write a Python Program to find the largest of three numbers.
 
-num1 = int(input("Enter the first number >>> "))
-num2 = int(input("Enter the second number >>> "))
-num3 = int(input("Enter the third number >>> "))
+num1 = float(input("Enter the first number >>> "))
+num2 = float(input("Enter the second number >>> "))
+num3 = float(input("Enter the third number >>> "))
 
 print(f"Largest of {num1, num2, num3} is {max(num1, num2, num3)}")
