+++ sec_b_q_2.py
@@ -3,5 +3,5 @@
 while True:
     num = input("Enter any number to check the divisibility by 7 >>> ")
     if not num: break
-    num = int(num)
+    num = float(num)
     print(f"Yes, {num} is divisible by 7" if num % 7 == 0 else f"No, {num} is not divisible by 7")
